+++ packages/expo-ui/ios/Modifiers/View+ModifierArray.swift
@@ -11,14 +11,20 @@ internal extension View {
    * Applies an array of modifiers to a view using the ViewModifierRegistry.
    */
   @ViewBuilder
-  func applyModifiers(_ modifiers: ModifierArray?, globalEventDispatcher: EventDispatcher) -> some View {
-    if let modifiers = modifiers {
+  func applyModifiers(_ modifiers: ModifierArray?, appContext: AppContext?, globalEventDispatcher: EventDispatcher) -> some View {
+    if let modifiers, let appContext {
       modifiers.reduce(AnyView(self)) { currentView, modifierConfig in
         guard let type = modifierConfig["$type"] as? String else {
           return currentView
         }
 
-        return ViewModifierRegistry.shared.applyModifier(type, to: currentView, globalEventDispatcher: globalEventDispatcher, params: modifierConfig)
+        return ViewModifierRegistry.shared.applyModifier(
+          type,
+          to: currentView,
+          appContext: appContext,
+          globalEventDispatcher: globalEventDispatcher,
+          params: modifierConfig
+        )
       }
     } else {
       self